Don't demand Jason
