+++ lib/joken/application.ex
@@ -9,25 +9,11 @@ defmodule Joken.Application do
 
     children = []
 
-    JOSE.json_module(JOSE.Jason)
+    if Code.ensure_loaded?(Jason) do
+      JOSE.json_module(JOSE.Jason)
+    end
+
     opts = [strategy: :one_for_one]
     Supervisor.start_link(children, opts)
   end
-end
-
-defmodule JOSE.Jason do
-  @moduledoc """
-  Implementation of the `:jose_json` behaviour for the Jason library.
-
-  It only delegates both `encode/1` and `decode/1` functions to Jason's calls.
-
-  Future work might want to encode/decode specific types like Poison's implementation.
-  """
-  @behaviour :jose_json
-
-  @impl true
-  def encode(term), do: Jason.encode!(term)
-
-  @impl true
-  def decode(binary) when is_binary(binary), do: Jason.decode!(binary)
 end
